Doctor's Login
Develop a Swing application which can keep record of doctor appointments. Patients are
supposed to submit their complaints (Cold, cough, fever, Headache, Bodypains, Vomit,
Nausea, Dehydration etc) and depending on the severity of the patients, the order of the
patient appointments get reordered.
Create a JFrame which allows the users to submit their details, which are saved in a file.
When the doctor logs into the application, the patients list is sorted according to the severity
of the patients disease which is displayed in a JList.